# docs/swagger.yaml
basePath: /api/v1
definitions:
  comments.Comment:
    properties:
      author_id:
        type: integer
      content:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      likes:
        description: Метаданные
        type: integer
      parent:
        allOf:
        - $ref: '#/definitions/comments.Comment'
        description: Древовидная структура
      parent_id:
        description: Для древовидной структуры
        type: integer
      post_id:
        type: integer
      replies:
        items:
          $ref: '#/definitions/comments.Comment'
        type: array
      status:
        $ref: '#/definitions/comments.Status'
      updated_at:
        type: string
    type: object
  comments.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      details:
        example: Поле 'title' не может быть пустым
        type: string
      message:
        example: Неверный формат данных
        type: string
    type: object
  comments.Status:
    enum:
    - active
    - deleted
    - hidden
    type: string
    x-enum-varnames:
    - StatusActive
    - StatusDeleted
    - StatusHidden
  posts.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      details:
        example: Поле 'title' не может быть пустым
        type: string
      message:
        example: Неверный формат данных
        type: string
    type: object
  posts.Post:
    properties:
      author_id:
        description: Связи
        type: integer
      comments:
        items:
          $ref: '#/definitions/comments.Comment'
        type: array
      created_at:
        description: Временные метки
        type: string
      description:
        type: string
      html_content:
        description: Отрендеренный HTML
        type: string
      id:
        type: integer
      published_at:
        type: string
      raw_content:
        description: Контент
        type: string
      slug:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/posts.Status'
        description: Метаданные
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
      view_count:
        type: integer
    type: object
  posts.Status:
    enum:
    - draft
    - published
    - archived
    type: string
    x-enum-varnames:
    - StatusDraft
    - StatusPublished
    - StatusArchived
host: localhost:8080
info:
  contact: {}
  description: API сервер для блог-платформы
  title: Blog Service API
  version: "1.0"
paths:
  /api/v1/posts:
    get:
      parameters:
      - description: Смещение
        in: query
        name: offset
        type: integer
      - description: Количество записей
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/posts.Post'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
      summary: Получить список постов
      tags:
      - posts
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные поста
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/posts.Post'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/posts.Post'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
      security:
      - JWT: []
      summary: Создать новый пост
      tags:
      - posts
  /api/v1/posts/{id}:
    delete:
      parameters:
      - description: ID поста
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
      security:
      - JWT: []
      summary: Удалить пост
      tags:
      - posts
    get:
      parameters:
      - description: ID поста
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/posts.Post'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
      summary: Получить пост по ID
      tags:
      - posts
    put:
      consumes:
      - application/json
      parameters:
      - description: ID поста
        in: path
        name: id
        required: true
        type: integer
      - description: Данные поста
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/posts.Post'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/posts.Post'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
      security:
      - JWT: []
      summary: Обновить пост
      tags:
      - posts
  /api/v1/posts/slug/{slug}:
    get:
      parameters:
      - description: Slug поста
        in: path
        name: slug
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/posts.Post'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
      summary: Получить пост по slug
      tags:
      - posts
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"